==> codes/02-CalculateAndStoreHashes.ipynb <==

import cv2
import os
import image_similarity_measures
from sys import argv
from image_similarity_measures.quality_metrics import rmse, ssim, sre
import numpy as np
import pandas as pd
import csv
import datetime
# get_ipython().run_line_magic('matplotlib', 'inline')
#The line above is necesary to show Matplotlib's plots inside a Jupyter Notebook
from matplotlib import pyplot as plt
import pickle

import imagehash
from PIL import Image


reference_locations=[]
path_to_ref="/ddn/gfxhome/asislam25/projects/other_misc_projects/video_copy/data/references/"
path_to_ref_kf="/ddn/gfxhome/asislam25/projects/other_misc_projects/video_copy/intermediate/keyframes/"
for f in os.listdir(path_to_ref):
    f=f.split(".")[0]
    reference_locations.append(path_to_ref_kf+f)
    
print(reference_locations,len(reference_locations))


intermediate_location="../intermediate/"
if not os.path.isdir(intermediate_location):
    os.mkdir(intermediate_location)
hash_location=intermediate_location+"hash_values/"
if not os.path.isdir(hash_location):
    os.mkdir(resized_location)

hash_location=hash_location+"phash/"
if not os.path.isdir(hash_location):
    os.mkdir(resized_location)
    
dic_location=hash_location+"dic_vals.p"    

if not os.path.isfile(dic_location):
    dic_hash_vals={}
else:
    dic_hash_vals = pickle.load( open( dic_location, "rb" ) )









len(dic_hash_vals.keys())


for reference_location_path in reference_locations:
    ref_file_frames_list=os.listdir(reference_location_path)
    file_name=reference_location_path.split("/")[-1]
    if file_name in dic_hash_vals:
        print("hash already generated for ",file_name)
        continue
    dic_hash_vals[file_name]={}
#     print(reference_location_path,file_name,ref_file_frames_list)
    for frame_name in ref_file_frames_list:
        f=os.path.join(reference_location_path,frame_name)
        hash_val=imagehash.phash(Image.open(f))
#         print(f,hash_val)
        dic_hash_vals[file_name][frame_name]=hash_val

dic_hash_vals.keys()

# dic_hash_vals

archive_locations=[]
path_to_archive="/ddn/gfxhome/asislam25/projects/other_misc_projects/video_copy/data/setC/positive-c.zip/"
path_to_archive_kf="/ddn/gfxhome/asislam25/projects/other_misc_projects/video_copy/intermediate/keyframes/"
for f in os.listdir(path_to_archive):
    f=f.split(".")[0]

    archive_locations.append(path_to_archive_kf+f)
    
print(archive_locations,len(archive_locations))    

for archive_location_path in archive_locations:
    arch_file_frames_list=os.listdir(archive_location_path)
    file_name=archive_location_path.split("/")[-1]
    if file_name in dic_hash_vals:
        print("hash already generated for ",file_name)
        continue    
    dic_hash_vals[file_name]={}
#     print(archive_location_path,file_name,arch_file_frames_list)
    print(archive_location_path,file_name)
    for frame_name in arch_file_frames_list:
        f=os.path.join(archive_location_path,frame_name)
        hash_val=imagehash.phash(Image.open(f))
#         print(f,hash_val)
        dic_hash_vals[file_name][frame_name]=hash_val

# intermediate_location="../intermediate/"
# if not os.path.isdir(intermediate_location):
#     os.mkdir(intermediate_location)
# hash_location=intermediate_location+"hash_values/"
# if not os.path.isdir(hash_location):
#     os.mkdir(resized_location)

# hash_location=hash_location+"phash/"
# if not os.path.isdir(hash_location):
#     os.mkdir(resized_location)
    
# dic_location=hash_location+"dic_vals.p"    
pickle.dump( dic_hash_vals, open( dic_location, "wb" ) )

len(dic_hash_vals.keys())
