==> codes/04-accuracyCalc.ipynb <==
import pandas as pd
import os
import numpy as np
from sklearn.metrics import accuracy_score

'''
in order to compare accuracy of our algorithm with that of
the ground truth, we need two inputs

input1=list of reference videos
input2=list of archive videos


'''

reference_locations=[]
path_to_ref="/ddn/gfxhome/asislam25/projects/other_misc_projects/video_copy/data/references/"
path_to_ref_kf="/ddn/gfxhome/asislam25/projects/other_misc_projects/video_copy/intermediate/keyframes/"
for f in os.listdir(path_to_ref):
    f=f.split(".")[0]
    reference_locations.append(path_to_ref_kf+f)
    
print(reference_locations,len(reference_locations))

archive_locations=[]
path_to_archive="/ddn/gfxhome/asislam25/projects/other_misc_projects/video_copy/data/setC/positive-c.zip/"
path_to_archive_kf="/ddn/gfxhome/asislam25/projects/other_misc_projects/video_copy/intermediate/keyframes/"
for f in os.listdir(path_to_archive):
    f=f.split(".")[0]

    archive_locations.append(path_to_archive_kf+f)
    
print(archive_locations,len(archive_locations))    

# # below are the two lists as inputs
# reference_locations=[
#     "/Users/ashhadulislam/projects/other_misc/video_copy/intermediate/keyframes/c01_202101300630",
#     "/Users/ashhadulislam/projects/other_misc/video_copy/intermediate/keyframes/c01_202101300810"    
# ]

# archive_locations=[
#     "/Users/ashhadulislam/projects/other_misc/video_copy/intermediate/keyframes/c01_20210130063213_C2",
#     "/Users/ashhadulislam/projects/other_misc/video_copy/intermediate/keyframes/c01_20210130081629_C3",    
#     "/Users/ashhadulislam/projects/other_misc/video_copy/intermediate/keyframes/c01_20210130115216_C3"
# ]


reference_location_video_names_only=[]
for full_path in reference_locations:
    just_name=full_path.split("/")[-1]
    reference_location_video_names_only.append(just_name)

archive_location_video_names_only=[]

for full_path in archive_locations:
    # remove everything before the last /
    just_name=full_path.split("/")[-1]
    
    # get till the last underscore
    just_name=just_name[:-3]
    archive_location_video_names_only.append(just_name)

print(reference_location_video_names_only,archive_location_video_names_only)
archive_location_video_names_only.sort()

len(reference_location_video_names_only),len(archive_location_video_names_only)

df=pd.read_csv("../data/groundtruth.csv",sep=";")

df.head()

df["Reference_Video_name_only"]=[nm.split(".")[0] for nm in df["Reference_Video"]]

df["Positive_Video_name_only"]=[nm.split(".")[0] for nm in df["Positive_Video"]]

df.head()

df=df[["Reference_Video_name_only","Positive_Video_name_only"]]

df.head()

df.shape

# apply filter for reference videos
df_new=[]
for a_vid in reference_location_video_names_only:
    df_chosen=df[df["Reference_Video_name_only"]==a_vid]
    print(a_vid,df_chosen.shape)
    df_new.append(df_chosen)

df_new=pd.concat(df_new)

df_new.shape

my_ground_truths={}
my_ground_truths["reference_video"]=[]
my_ground_truths["score_array"]=[]
my_ground_truths["archive_videos"]=[]

for a_vid in reference_location_video_names_only:
    df_chosen=df_new[df_new["Reference_Video_name_only"]==a_vid]
    print(df_chosen.shape)
    uniq_archives=df_chosen["Positive_Video_name_only"].unique()
    print(uniq_archives)
    scores=[]
    for arch_vid in archive_location_video_names_only:
        if arch_vid in uniq_archives:
            scores.append(1)
        else:
            scores.append(0)
    print(scores)
    my_ground_truths["reference_video"].append(a_vid)
    my_ground_truths["score_array"].append(scores)
    my_ground_truths["archive_videos"].append(archive_location_video_names_only)

df_gtruth=pd.DataFrame(my_ground_truths)
print(df_gtruth.shape)

df_gtruth.head()

df_gtruth.to_csv("../results/computer_says.csv",index=False)

## now we need to process our output

# choose a strategy and threshold
strategy="phash"
threshold=10

df_res_us=pd.read_csv("../results/res_all.csv")
print(df_res_us.shape)
df_res_us=df_res_us[df_res_us["strategy"]==strategy]
df_res_us=df_res_us[df_res_us["score"]<=threshold]

df_res_us.shape

df_res_us.head()

df_res_us["ref_vid_name_only"]=[nm.split("/")[-2] for nm in df_res_us["ref_kf_path_name"]]


df_res_us["arch_vid_name_only"]=[full_path.split("/")[-1][:-3] for full_path in df_res_us["arch_kf_path"]]

df_res_us=df_res_us[["ref_vid_name_only","arch_vid_name_only"]]

df_res_us.head()

# now to calculate the scores
my_calculations={}
my_calculations["reference_video"]=[]
my_calculations["score_array"]=[]
my_calculations["archive_videos"]=[]

for a_vid in reference_location_video_names_only:
    df_chosen=df_res_us[df_res_us["ref_vid_name_only"]==a_vid]
    print(df_chosen.shape)
    uniq_archives=df_chosen["arch_vid_name_only"].unique()
    print(uniq_archives)
    scores=[]
    for arch_vid in archive_location_video_names_only:
        if arch_vid in uniq_archives:
            scores.append(1)
        else:
            scores.append(0)
    print(scores)
    my_calculations["reference_video"].append(a_vid)
    my_calculations["score_array"].append(scores)
    my_calculations["archive_videos"].append(archive_location_video_names_only)

df_my_res=pd.DataFrame(my_calculations)

df_my_res.head()

df_my_res.to_csv("../results/our_calcul.csv",index=False)

### Compare and calculate the scores

df_given=pd.read_csv("../results/computer_says.csv")
df_calculated=pd.read_csv("../results/our_calcul.csv")

reference_vids=list(df_given.reference_video)
avg_acc=0
for vid in reference_vids:
    df_given_vid=df_given[df_given["reference_video"]==vid]
    df_calculated_vid=df_calculated[df_calculated["reference_video"]==vid]
#     print(df_given_vid.shape,df_calculated_vid.shape)

    comp_array=list(df_given_vid["score_array"])[0].split(",")
    comp_array=comp_array[1:-2]    
    comp_array=[int(v) for v in comp_array]
    comp_array=np.asarray(comp_array)
    
    calc_array=list(df_calculated_vid["score_array"])[0].split(",")
    calc_array=calc_array[1:-2]    
    calc_array=[int(v) for v in calc_array]
    calc_array=np.asarray(calc_array)

    acc=accuracy_score(calc_array,comp_array)
    avg_acc+=acc
    print(vid,acc)
avg_acc=avg_acc/len(reference_vids)

print("Average accuracy is ",avg_acc)
